startup latency and resource efficiency/resourcea/benchmark/python/ml/serving/cnn_image_classification/azure/requirements.txt
azure-functions==1.0.0b3
azure_storage_blob==1.0.0
tensorflow==1.14.0
Pillow